requirements.txt
googleapis-common-protos==1.53.0
grpcio==1.40.0
protobuf==3.18.0
six==1.16.0
yachalk==0.1.4
flask==2.0.1
python-dotenv==0.19.0